fix: add BaseStrokeThickness property to IMapAnchorViewModel
This commit introduces property for the initial stroke thickness of map anchors: "BaseStrokeThickness." This property allows configuring visual appearance of map anchors for better fit and precision. It can be adjusted independently and will affect the rendering of the map anchors on the map.

Asana: https://app.asana.com/0/1203851531040615/1207177984234823/f